Backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
pymongo==4.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
